requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
google-genai==1.0.0
python-dotenv==1.0.0
pyotp==2.9.0
